% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix_helper.R
\name{addGeneNames}
\alias{addGeneNames}
\title{The addGeneNames function}
\usage{
addGeneNames(df_data, ProbeAnnotation)
}
\arguments{
\item{df_data}{a dataframe with a column named Probe}

\item{ProbeAnnotation}{a dataframe with ProbeAnnotation, including one column named "probe" and another column named "gene"}
}
\value{
a dataframe with added gene names
}
\description{
Given a dataframe with a column of probe names, add the gene names
}
