% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{BatchCorrection_Combat}
\alias{BatchCorrection_Combat}
\title{The BatchCorrection_Combat function}
\usage{
BatchCorrection_Combat(GEN_Data, BatchDataSelected)
}
\arguments{
\item{GEN_Data}{matrix with methylation.data or gene.expression.data}

\item{BatchDataSelected}{BatchData after filtering out the small batches and selecting for overlapped samples}
}
\value{
corrected data matrix
}
\description{
The BatchCorrection_Combat function
}
\details{
correct batch effects with Combat
}
\keyword{internal}
