% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{BatchCorrection_Seurat}
\alias{BatchCorrection_Seurat}
\title{The BatchCorrection_Seurat function}
\usage{
BatchCorrection_Seurat(GEN_Data, BatchDataSelected)
}
\arguments{
\item{GEN_Data}{matrix with methylation.data or gene.expression.data}

\item{BatchDataSelected}{BatchData after filtering out the small batches and selecting for overlapped samples.}
}
\value{
corrected data matrix
}
\description{
The BatchCorrection_Seurat function
}
\details{
correct batch effects with the Seurat data integration functions.
}
\keyword{internal}
