% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNANetwork.R
\name{.extractPriMiRNA}
\alias{.extractPriMiRNA}
\title{The extractPriMiRNA function}
\usage{
.extractPriMiRNA(str)
}
\arguments{
\item{str}{a character string for a mature miRNA name (e.g. "hsa-miR-34a-3p")}
}
\value{
a character string for the corresponding pri-miRNA name (e.g. "hsa-mir-34a")
}
\description{
Utility function to convert mature miRNA names to pri-miRNA names
}
