% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_gene_expression.R
\name{.getComp}
\alias{.getComp}
\title{The .getComp function}
\usage{
.getComp(state)
}
\arguments{
\item{state}{character string indicating the methylation state, can be either "Hyper", "Hypo", "Dual"}
}
\value{
a list of sample names split by methylation group
}
\description{
Helper function to get a string indicating the comparison made for gene expression
}
\keyword{internal}
