% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_gene_expression.R
\name{.getMetGroup}
\alias{.getMetGroup}
\title{The .getMetGroup function}
\usage{
.getMetGroup(state, DM_values)
}
\arguments{
\item{state}{character string indicating the methylation state, can be either "Hyper", "Hypo", "Dual"}

\item{DM_values}{a vector of DM values for the probe. The names of the vector are sample names.}
}
\value{
a list of sample names split by methylation group
}
\description{
Helper function to get sample names split by methylation group based on DM values
}
\keyword{internal}
