% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.mapProbeGene}
\alias{.mapProbeGene}
\title{The .mapProbeGene function}
\usage{
.mapProbeGene(df.annot)
}
\arguments{
\item{df.annot}{a dataframe with probe annotation, can be the object returned from the convertAnnotToDF function.}
}
\value{
a dataframe with 1:1 mapping of probe and gene
}
\description{
since in the original probe annotation, a specific probe can be mapped to multiple genes, this function splits the rows and maps each probe to a signle gene in a row.
}
\keyword{internal}
