% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.splitMetData}
\alias{.splitMetData}
\title{The .splitMetData function}
\usage{
.splitMetData(methylation.data, sample.info, group.1, group.2)
}
\arguments{
\item{methylation.data}{methylation data matrix}

\item{sample.info}{sample information matrix}

\item{group.1}{name of group.1}

\item{group.2}{name of group.2}
}
\value{
a list with methylation data of group.1 and group.2
}
\description{
Helper function to split the methylation data matrix into the experimental group and the control group
}
\keyword{internal}
