% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{EpiMix_getInfiniumAnnotation}
\alias{EpiMix_getInfiniumAnnotation}
\title{The EpiMix_getInfiniumAnnotation function}
\usage{
EpiMix_getInfiniumAnnotation(plat = "EPIC", genome = "hg38")
}
\arguments{
\item{plat}{character string indicating the methylation platform.}

\item{genome}{character string indicating the version of genome build}
}
\value{
a GRange object of probe annotation
}
\description{
fetch the Infinium probe annotation from the AnnotationHub
}
\examples{
\donttest{
annot <- EpiMix_getInfiniumAnnotation(plat = "EPIC", genome = "hg38")
}
}
\keyword{annotation}
