% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{EpiMix_ModelGeneExpression}
\alias{EpiMix_ModelGeneExpression}
\title{The EpiMix_ModelGeneExpression function}
\usage{
EpiMix_ModelGeneExpression(
  methylation.data,
  gene.expression.data,
  ProbeAnnotation,
  cores,
  filter
)
}
\arguments{
\item{methylation.data}{methylation data matrix.}

\item{gene.expression.data}{gene expression data matrix.}

\item{ProbeAnnotation}{dataframe of probe annotation}

\item{cores}{number of CPU cores used for computation}

\item{filter}{logical indicating whether to perform a linear regression to select functional probes}
}
\value{
a character vector of probe names.
}
\description{
pre-select funcitonal probes.
}
\keyword{internal}
