% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_gene_expression.R
\name{filterLinearProbes}
\alias{filterLinearProbes}
\title{The filterLinearProbes function}
\usage{
filterLinearProbes(
  methylation.data,
  gene.expression.data,
  ProbeAnnotation,
  cores,
  filter,
  cluster,
  correlation = "negative"
)
}
\arguments{
\item{methylation.data}{methylation data matrix.}

\item{gene.expression.data}{gene expression data matrix.}

\item{ProbeAnnotation}{dataframe of probe annotation}

\item{cores}{number of CPU cores used for computation}

\item{filter}{logical indicating whether to perform a linear regression to select functional probes}

\item{cluster}{logical indicating whether the CpGs were clustered using hierarchical clustering}

\item{correlation}{Character vector indicating the expected correlation between DNA methylation and gene expression. Can be either 'negative' or 'positive'. Default: 'negative'.}
}
\value{
a character vector of probe names.
}
\description{
use a linear regression filter to screen for probes that were negatively associated with gene expression.
}
\keyword{internal}
