% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filterMethMatrix}
\alias{filterMethMatrix}
\title{The filterMethMatrix function}
\usage{
filterMethMatrix(MET_matrix, gene.expression.data)
}
\arguments{
\item{MET_matrix}{a matrix of methylation states from the EpiMix results}

\item{gene.expression.data}{a matrix with gene expression data}
}
\value{
a matrix of methylation states for each differentially methylated probe with probes in rows and patient in columns.
}
\description{
The filterMethMatrix function
}
\details{
This function filters methylation states from the beta mixture modeling for each probe. The filtered probes can be used to model gene expression by Wilcoxon test.
}
\keyword{internal}
