% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMix_helper.R
\name{filterProbes}
\alias{filterProbes}
\title{The filterProbes function}
\usage{
filterProbes(
  mode,
  gene.expression.data,
  listOfGenes,
  promoters,
  met.platform,
  genome
)
}
\arguments{
\item{mode}{analytic mode}

\item{gene.expression.data}{matrix of gene expression data}

\item{listOfGenes}{list of genes of interest}

\item{promoters}{logic indicating whether to filter CpGs on promoters}

\item{met.platform}{methylation platform}

\item{genome}{genome build version}
}
\value{
filtered ProbeAnnotation
}
\description{
filter CpG sites based on user-specified conditions
}
