% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_gene_expression.R
\name{generateFunctionalPairs}
\alias{generateFunctionalPairs}
\title{The generateFunctionalPairs function}
\usage{
generateFunctionalPairs(
  MET_matrix,
  control.names,
  gene.expression.data,
  ProbeAnnotation,
  raw.pvalue.threshold,
  adjusted.pvalue.threshold,
  cores,
  mode = "Regular",
  correlation = "negative"
)
}
\arguments{
\item{MET_matrix}{matrix of methylation states}

\item{control.names}{character vector indicating the samples names in the control group}

\item{gene.expression.data}{matrix of gene expression data}

\item{ProbeAnnotation}{dataframe of probe annotation}

\item{raw.pvalue.threshold}{raw p value threshold}

\item{adjusted.pvalue.threshold}{adjusted p value threshold}

\item{cores}{number of computational cores}

\item{mode}{character string indicating the analytic mode}

\item{correlation}{the expected relationship between DNAme and gene expression}
}
\value{
a dataframe of functional CpG-gene matrix
}
\description{
Wrapper function to get functional CpG-gene pairs, used for Regular, miRNA and lncRNA modes
}
