% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_Download_Preprocess.R
\name{GEO_EstimateMissingValues_Molecular}
\alias{GEO_EstimateMissingValues_Molecular}
\title{The GEO_EstimateMissingValues_Molecular function}
\usage{
GEO_EstimateMissingValues_Molecular(
  MET_Data,
  MissingValueThresholdGene = 0.3,
  MissingValueThresholdSample = 0.1
)
}
\arguments{
\item{MET_Data}{methylation data or gene expression data matrix.}

\item{MissingValueThresholdGene}{threshold for missing values per gene. Genes with a percentage of NAs greater than this threshold are removed. Default is 0.3.}

\item{MissingValueThresholdSample}{threshold for missing values per sample. Samples with a percentage of NAs greater than this threshold are removed. Default is 0.1.}
}
\value{
the dataset with imputed values and possibly some genes or samples deleted.
}
\description{
Internal. Removes samples and genes with more missing values than the MissingValueThreshold, and imputes remaining missing values using Tibshirani's KNN method.
}
\keyword{internal}
