% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_Download_Preprocess.R
\name{GEO_GetSampleInfo}
\alias{GEO_GetSampleInfo}
\title{The GEO_GetSampleInfo function}
\usage{
GEO_GetSampleInfo(METdirectories, group.column, targetDirectory = ".")
}
\arguments{
\item{METdirectories}{list of the file paths to the downloaded DNA methylation data, which can be the output from the GEO_Download_DNAMethylation function.}

\item{group.column}{character string indicating the column in the phenotypic data that defines the study group of each sample. The values in this column will be used to split the experiment and the control group.}

\item{targetDirectory}{file path to save the output. Default: '.' (current directory)}
}
\value{
a dataframe with two columns: a 'primary' column indicating the actual sample names, a 'sample.type' column indicating the study group for each sample.
}
\description{
auxiliary function to generate a sample information dataframe that indicates which study group each sample belongs to.
}
\keyword{purpose}
\keyword{testing}
