% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_Download_Preprocess.R
\name{GEO_getSampleMap}
\alias{GEO_getSampleMap}
\title{the GEO_getSampleMap function}
\usage{
GEO_getSampleMap(METdirectories, GEdirectories, targetDirectory = ".")
}
\arguments{
\item{METdirectories}{list of the file paths to the downloaded DNA methylation datasets, which can be the output from the GEO_Download_DNAMethylation function.}

\item{GEdirectories}{list of the file paths to the downloaded gene expression datasets, which can be the output from the GEO_Download_GeneExpression function.}

\item{targetDirectory}{file path to save the output. Default: '.' (current directory)}
}
\value{
dataframe with three columns: $assay (character string indicating the type of the experiment, can be either 'DNA methylation' or 'Gene expression'), $primary(character string indicating the actual sample names), $colnames (character string indicating the actual column names for each samples in DNA methylation data and gene expression data)
}
\description{
auxiliary function to generate a sample map for DNA methylation data and gene expression data
}
\keyword{purpose}
\keyword{testing}
