% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get.chromosome}
\alias{get.chromosome}
\title{The get.chromosome function}
\usage{
get.chromosome(genes, genome)
}
\arguments{
\item{genes}{character vector with the gene names}

\item{genome}{character string indicating the genome build version, can be either "hg19" or "hg38"}
}
\value{
a dataframe for the mapping between genes and their chromosomes.
}
\description{
given a list of genes, get the chromosomes of these genes.
}
\keyword{internal}
