% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get.comparators}
\alias{get.comparators}
\title{The get.comparators function}
\usage{
get.comparators(state)
}
\arguments{
\item{state}{the methylation state of a CpG site. Can be either 'Hyper', 'Hypo' or 'Dual'.}
}
\value{
a string indicating the comparisions for gene expression.
}
\description{
Helper function to generate the string indicating the comparisons of gene expression. Used for generating the functional CpG-gene pair matrix.
}
\keyword{internal}
