% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get.perm.pVals}
\alias{get.perm.pVals}
\title{The get.perm.pVals function}
\usage{
get.perm.pVals(
  target.probe,
  MET_matrix,
  gene.expression.data,
  ProbeAnnotation,
  genome = "hg38",
  correlation = "negative",
  perm = 1000
)
}
\arguments{
\item{target.probe}{character string indicating the target CpG for generating the permutation p values.}

\item{MET_matrix}{a matrix of methylaiton states for the patients.}

\item{gene.expression.data}{a matrix of gene expression data.}

\item{ProbeAnnotation}{GRange object of probe annotation.}

\item{genome}{character string indicating the genome build version, can be either "hg19" or "hg38".}

\item{correlation}{the direction of correlation between the methylation state of the CpG site and the gene expression levels. Can be either "negative" or "positive".}

\item{perm}{the number of permutation tests.}
}
\value{
a dataframe for the permutation genes and p values for the target CpG site.
}
\description{
Helper function to perform permutation test for a target enhancer CpG. Used in the Enhancer mode.
}
\keyword{internal}
