% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get.prevalence}
\alias{get.prevalence}
\title{The get.prevalence function}
\usage{
get.prevalence(MethylMixResults)
}
\arguments{
\item{MET_matrix}{matrix of methylation states}
}
\value{
a list of prevalence for the abnormal methylation
}
\description{
Helper function to get the methylation state and the
prevalence of the differential methylation of a CpG sites in the study population.
}
\keyword{internal}
