% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get.survival.probe}
\alias{get.survival.probe}
\title{The get.survival.probe function}
\usage{
get.survival.probe(
  EpiMixResults,
  TCGA_CancerSite = NULL,
  clinical.data = NULL,
  pval.threshold = 0.05,
  OutputRoot = ""
)
}
\arguments{
\item{EpiMixResults}{List of objects returned from the EpiMix function}

\item{TCGA_CancerSite}{TCGA cancer code (e.g. "LUAD")}

\item{clinical.data}{(If the TCGA_CancerSite parameter has been specified, this parameter is optional) Dataframe with survival information. Must contain at least three columns: "sample.id", "days_to_death", "days_to_last_follow_up".}

\item{pval.threshold}{numeric value indicting the p value threshold for selecting the survival predictive probes. Survival time is compared by log-rank test. Default: 0.05}

\item{OutputRoot}{path to save the output. If not null, the return value will be saved as "Survival)Probes.csv".}
}
\value{
a dataframe with probes whose methylation state is predictive of patient survival and the p value.
}
\description{
Get probes whose methylation state is predictive of patient survival
}
\examples{
\dontrun{
library(survival)

data("Sample_EpiMixResults_miRNA")

# Set the TCGA cancer site.
CancerSite = "LUAD"

# Find survival-associated CpGs/genes
survival.CpGs <- get.survival.probe (EpiMixResults = Sample_EpiMixResults_miRNA, TCGA_CancerSite = CancerSite)
}

}
