% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getFunctionalProbes}
\alias{getFunctionalProbes}
\title{The getFunctionalProbes function}
\usage{
getFunctionalProbes(
  gene,
  probes,
  MET_matrix,
  MET_Control,
  exp,
  methylation.states,
  raw.pvalue.threshold = 0.05,
  adjusted.pvalue.threshold = 0.01
)
}
\arguments{
\item{gene}{character string indicating the target gene to be modeled.}

\item{probes}{character vector indicating the probes mapped to the target gene.}

\item{MET_matrix}{methylation data matrix for CpGs from group.1 and group.2.}

\item{MET_Control}{methylation data matrix for CpGs from group.2.}

\item{exp}{gene expression data matrix.}

\item{methylation.states}{character vector indicating the methylation states of the target probes}

\item{raw.pvalue.threshold}{raw p value from testing DNA methylation and gene expression}

\item{adjusted.pvalue.threshold}{adjusted p value from testing DNA methylation and gene expression}
}
\value{
dataframe with functional probe-gene pairs and corresponding p values obtained from the Wilcoxon test for gene expression and methylation.
}
\description{
Helper function to assess if the expression of a specific gene is reversely correlated with the methylation of a list of probes mapped to it.
}
\details{
This function is gene-centered, which is used in the regular mode and the lncRNA mode of EpiMix.
}
\keyword{internal}
