% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{getLncRNAData}
\alias{getLncRNAData}
\title{The getLncRNAData function}
\usage{
getLncRNAData(CancerSite)
}
\arguments{
\item{CancerSite}{TCGA cancer code}
}
\value{
local file path where the lncRNA expression data are saved
}
\description{
Helper function to retrieve the lncRNA expression data from Experiment Hub
}
\keyword{internal}
