% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getMethStates}
\alias{getMethStates}
\title{The getMethStates function}
\usage{
getMethStates(MethylMixResults, DM.probes)
}
\arguments{
\item{MethylMixResults}{the list object returned from the EpiMix function}

\item{DM.probes}{character vector of differentially methylated probes.}
}
\value{
a character vector with the methylation state ('Hypo', 'Hyper' or 'Dual') for each probe. The names for the vector are the probe names and the values are the methylation state.
}
\description{
Helper function that adds a methyaltion state label to each driver probe
}
\keyword{internal}
