% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getMethStates_Helper}
\alias{getMethStates_Helper}
\title{The getMethStates_Helper function}
\usage{
getMethStates_Helper(DMValues)
}
\arguments{
\item{DMValues}{a character vector indicating the DM values of a CpG site}
}
\value{
a character string incdicating the methylation state of the CpG
}
\description{
helper function to determine the methylation state based on DM values
}
