% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getRandomGenes}
\alias{getRandomGenes}
\title{The getRandomGenes function}
\usage{
getRandomGenes(
  target.probe,
  gene.expression.data,
  ProbeAnnotation,
  genome = "hg38",
  perm = 1000
)
}
\arguments{
\item{target.probe}{character string indicating the target CpG for generating the permutation p values.}

\item{gene.expression.data}{a matrix of gene expression data.}

\item{ProbeAnnotation}{GRange object of probe annotation.}

\item{genome}{character string indicating the genome build version, can be either 'hg19' or 'hg38'.}

\item{perm}{the number of permutation tests. Default: 1000}
}
\value{
a dataframe for the permutation genes and p values for the target CpG site.
}
\description{
Helper function to get a set of random genes located on different chromosomes of the target CpG.
}
\keyword{internal}
