% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mapTranscriptToGene}
\alias{mapTranscriptToGene}
\title{mapTranscriptToGene}
\usage{
mapTranscriptToGene(transcripts)
}
\arguments{
\item{transcripts}{vector with the name of miRNA precursors}
}
\value{
a dataframe with two columns: "Transcript" indicating the miRNA precursor names, "Gene_name" indicating the actual human gene names (HGNC)
}
\description{
map the miRNA precursor names to HGNC
}
\keyword{internal}
