% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMix.R
\name{MethylMix_Predict}
\alias{MethylMix_Predict}
\title{The MethylMix_Predict function}
\usage{
MethylMix_Predict(newBetaValuesMatrix, MethylMixResult)
}
\arguments{
\item{newBetaValuesMatrix}{Matrix with new observations for prediction,
genes/cpg sites in rows, samples in columns. Although this new matrix can have
a different number of genes/cpg sites than the one provided as METcancer when
running MethylMix, naming of genes/cpg sites should be the same.}

\item{MethylMixResult}{Output object from MethylMix}
}
\value{
A matrix with predictions (indices of mixture component), driver genes in rows, new samples in columns
}
\description{
Given a new data set with methylation data, this function predicts the mixture
component for each new sample and driver gene. Predictions are based on posterior
probabilities calculated with MethylMix'x fitted mixture model.
}
