% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{Preprocess_MAdata_Normal}
\alias{Preprocess_MAdata_Normal}
\title{The Preprocess_MAdata_Normal function}
\usage{
Preprocess_MAdata_Normal(
  CancerSite,
  Directory,
  File,
  MissingValueThresholdGene,
  MissingValueThresholdSample,
  doBatchCorrection,
  batch.correction.method,
  BatchData
)
}
\arguments{
\item{CancerSite}{TCGA code for the cancer site.}

\item{Directory}{Directory.}

\item{File}{File.}

\item{MissingValueThresholdGene}{threshold for missing values per gene. Genes with a percentage of NAs greater than this threshold are removed. Default is 0.3.}

\item{MissingValueThresholdSample}{threshold for missing values per sample. Samples with a percentage of NAs greater than this threshold are removed. Default is 0.1.}
}
\value{
The data matrix.
}
\description{
Internal. Pre-process gene expression data for normal samples.
}
\keyword{internal}
