% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO_Download_Preprocess.R
\name{removeDuplicatedGenes}
\alias{removeDuplicatedGenes}
\title{The removeDuplicatedGenes function}
\usage{
removeDuplicatedGenes(GEN_data)
}
\arguments{
\item{GEN_data}{gene expression data matrix}
}
\value{
gene expression data matrix with duplicated genes removed
}
\description{
sum up the transcript expression values if a gene has multiple transcripts
}
