% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{splitmatrix}
\alias{splitmatrix}
\title{The splitmatix function}
\usage{
splitmatrix(x, by = "row")
}
\arguments{
\item{x}{A matrix}

\item{by}{A character specify if split the matrix by row or column.}
}
\value{
A list each of which is the value of each row/column in the matrix.
}
\description{
The splitmatix function
}
\keyword{internal}
