% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{TCGA_Download_DNAmethylation}
\alias{TCGA_Download_DNAmethylation}
\title{The TCGA_Download_DNAmethylation function}
\usage{
TCGA_Download_DNAmethylation(CancerSite, TargetDirectory, downloadData = TRUE)
}
\arguments{
\item{CancerSite}{character of length 1 with TCGA cancer code.}

\item{TargetDirectory}{character with directory where a folder for downloaded files will be created.}

\item{downloadData}{logical indicating if data should be downloaded (default: TRUE). If false, the url of the desired data is returned.}
}
\value{
list with paths to downloaded files for both 27k and 450k methylation data.
}
\description{
Download DNA methylation data from TCGA.
}
\examples{
\dontrun{
CancerSite <- "OV"
targetDirctory <- paste0(getwd(), "/Data")
METdirectories <- TCGA_Download_DNAmethylation(CancerSite, targetDirectory)
}

}
\keyword{download}
