% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{TCGA_EstimateMissingValues_MolecularData}
\alias{TCGA_EstimateMissingValues_MolecularData}
\title{The TCGA_EstimateMissingValues_MolecularData function}
\usage{
TCGA_EstimateMissingValues_MolecularData(
  MET_Data,
  MissingValueThresholdGene = 0.3,
  MissingValueThresholdSample = 0.1
)
}
\arguments{
\item{MET_Data}{matrix of gene expression data}

\item{MissingValueThresholdGene}{threshold for missing values per gene. Genes with a percentage of NAs greater than this threshold are removed. Default is 0.3.}

\item{MissingValueThresholdSample}{threshold for missing values per sample. Samples with a percentage of NAs greater than this threshold are removed. Default is 0.1.}
}
\value{
gene expression data with no missing values.
}
\description{
Internal.Deletes samples and genes with more NAs than the respective thresholds. Imputes other NAs values.
}
\keyword{internal}
