% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_Download_Preprocess.R
\name{TCGA_Select_Dataset}
\alias{TCGA_Select_Dataset}
\title{The TCGA_Select_Dataset function}
\usage{
TCGA_Select_Dataset(CancerSite, MET_Data_27K, MET_Data_450K, use450K)
}
\arguments{
\item{CancerSite}{TCGA cancer code}

\item{MET_Data_27K}{matrix of MET_Data_27K}

\item{MET_Data_450K}{matrix of MET_Data_450K}

\item{use450K}{logic indicating whether to force use 450K data}
}
\value{
the selected MET data set
}
\description{
internal function to select which MET dataset to use
}
