% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGClustering.R
\name{translateMethylMixResults}
\alias{translateMethylMixResults}
\title{The translateMethylMixResults function}
\usage{
translateMethylMixResults(MethylMixResults, probeMapping)
}
\arguments{
\item{MethylMixResults}{list of MethylMix output}

\item{probeMapping}{dataframe of probe to gene-cluster mapping}
}
\value{
list of unfolded MethylMix results
}
\description{
unfold clustered MethylMix results to single CpGs
}
