% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validEpigenomes}
\alias{validEpigenomes}
\title{The validEpigenomes function}
\usage{
validEpigenomes(roadmap.epigenome.groups, roadmap.epigenome.ids)
}
\arguments{
\item{roadmap.epigenome.groups}{epigenome groups}

\item{roadmap.epigenome.ids}{epigenome ids}
}
\value{
a character vector of selected epigenome ids
}
\description{
check user input for roadmap epigenome groups or ids
}
