% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{model.gene.expression}
\alias{model.gene.expression}
\title{The model.gene.expression function}
\usage{
model.gene.expression(
  target.probe,
  MET_matrix,
  gene.expression.data,
  target.genes,
  state,
  correlation = "negative"
)
}
\arguments{
\item{target.probe}{character string indicating the target CpG site.}

\item{MET_matrix}{combined methylation state matrix for both case and control groups.}

\item{gene.expression.data}{a matrix with gene expression data.}

\item{target.genes}{character vector with the names of the genes associated with the target CpGs.}

\item{state}{methylation state for the target CpG, can be "Hyper", "Hypo" or "Dual".}

\item{correlation}{the direction of correlation between the methylation state of the CpG site and the gene expression levels. Can be either "negative" or "positive".}
}
\value{
a dataframe with fold change of gene expression and p values.
}
\description{
The model.gene.expression function
}
\keyword{internal}
